+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- [integrations/vue] feat: Vue performance monitoring (#2571)
 - "You miss 100 percent of the chances you don't take. — Wayne Gretzky" — Michael Scott
 
 - [core] feat: Send transactions in envelopes (#2553)